refactor: split happy/unhappy scenarios in separate files
