+++ __tests__/routers/groupRouter/unhappy.test.ts
@@ -1,16 +1,16 @@
 import request from 'supertest';
-import { GroupServiceImpl } from '../../src/services/group/group.service';
-import { GroupRouter } from '../../src/routers/group.router';
+import { GroupServiceImpl } from '../../../src/services/group/group.service';
+import { GroupRouter } from '../../../src/routers/group.router';
 import { Router } from 'express';
-import { Group, InternalError, ErrorType } from '../../src/models';
+import { Group, InternalError, ErrorType } from '../../../src/models';
 import express from 'express';
 import { v4 as uuidv4 } from 'uuid';
-import { internalErrorMidleware, defaultErrorMiddleware } from '../../src/middlewares';
-import { loggerService } from '../../src/configs/logger';
-import GroupService from '../../src/services/group/group.interface';
+import { internalErrorMidleware, defaultErrorMiddleware } from '../../../src/middlewares';
+import { loggerService } from '../../../src/configs/logger';
+import GroupService from '../../../src/services/group/group.interface';
 
-jest.mock('../../src/services/group/group.service.ts');
-jest.mock('../../src/configs/logger');
+jest.mock('../../../src/services/group/group.service.ts');
+jest.mock('../../../src/configs/logger');
 
 describe('Group router instance', () => {
   const id = uuidv4();
@@ -40,18 +40,6 @@ describe('Group router instance', () => {
   });
 
   describe('when get by id is called', () => {
-    test('should return group model', async () => {
-      //given
-      groupService.getById.mockImplementationOnce(() => Promise.resolve(group));
-      //when
-      await request(app)
-        .get('/' + id)
-        .expect(200)
-        .then((response) => expect(response.body).toEqual(group));
-      //then
-      expect(groupService.getById.mock.calls[0][0]).toEqual(id);
-    });
-
     test('should return error if invalid id specified', async () => {
       //given
       groupService.getById.mockImplementationOnce(() =>
@@ -68,19 +56,6 @@ describe('Group router instance', () => {
   });
 
   describe('when create group operation is called', () => {
-    test('should return created group', async () => {
-      //given
-      groupService.create.mockImplementationOnce(() => Promise.resolve(group));
-      //when
-      await request(app)
-        .post('/')
-        .send(group)
-        .expect(201)
-        .then((response) => expect(response.body).toEqual(group));
-      //then
-      expect(groupService.create.mock.calls[0][0]).toEqual(group);
-    });
-
     test('should return error if invalid permission specified', async () => {
       //given
       const invalidGroup = { ...group, permission: [...group.permissions, 'invalid'] };
@@ -143,19 +118,6 @@ describe('Group router instance', () => {
   });
 
   describe('when update group operation is called', () => {
-    test('should return updated group', async () => {
-      //given
-      groupService.update.mockImplementationOnce(() => Promise.resolve(group));
-      //when
-      await request(app)
-        .put('/' + id)
-        .send(group)
-        .expect(200)
-        .then((response) => expect(response.body).toEqual(group));
-      //then
-      expect(groupService.update.mock.calls[0][0]).toEqual(group);
-    });
-
     test('should return error if invalid permission specified', async () => {
       //given
       const invalidGroup = { ...group, permission: [...group.permissions, 'invalid'] };
@@ -210,51 +172,6 @@ describe('Group router instance', () => {
         );
       //then
       expect(groupService.update).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('when delete group operation is called', () => {
-    test('should remove group', async () => {
-      //given
-      groupService.delete.mockImplementationOnce(() => Promise.resolve());
-      //when
-      await request(app)
-        .delete('/' + id)
-        .expect(200)
-        .then((response) => expect(response.body).toEqual({}));
-      //then
-      expect(groupService.delete.mock.calls[0][0]).toEqual(id);
-    });
-  });
-
-  describe('when get all groups operation is called', () => {
-    test('should return list of groups', async () => {
-      //given
-      groupService.getAll.mockImplementationOnce(() => Promise.resolve([group]));
-      //when
-      await request(app)
-        .get('/')
-        .expect(200)
-        .then((response) => expect(response.body).toEqual([group]));
-      //then
-      expect(groupService.getAll).toHaveBeenCalled();
-    });
-  });
-
-  describe('when add users in group is called', () => {
-    test('should update users-group assotication', async () => {
-      //given
-      const userIds = [uuidv4(), uuidv4()];
-      groupService.updateUserGroupAssociation.mockImplementationOnce(() => Promise.resolve(group));
-      //when
-      await request(app)
-        .put(`/${id}/users`)
-        .send(userIds)
-        .expect(200)
-        .then((response) => expect(response.body).toEqual(group));
-      //then
-      expect(groupService.updateUserGroupAssociation.mock.calls[0][0]).toEqual(id);
-      expect(groupService.updateUserGroupAssociation.mock.calls[0][1]).toEqual(userIds);
     });
   });
 });
